Implement basic Ruby SDK for Kubera
Summary:
This is an initial implementation of the Ruby SDK for Kubera. It's still a work in progress and some of the code will probably change. Since the Kubera API is not live yet, it can't be completed (looking to use VCR to record actual requests) for now, but this patch could be merged first.

Update Rubocop config

Update configuration class

Implement Client, Request and Response

Add Base and User models

Test Plan:
1. `rspec`
2. `rubocop`

Reviewers: sirn, robin, unnawut

Reviewed By: sirn, robin, unnawut

Maniphest Tasks: T183

Differential Revision: https://phabricator.omisego.io/D42